Bcloud-3431

+++ src/DefaultRelayTCPSocket.cpp
@@ -160,7 +160,7 @@ namespace BrainCloud
             return;
         }
 
-        int ret = 0;
+        size_t ret = 0;
         while (ret >= 0 && ret < (int)size)
         {
             ret = ::send(m_socket, (const char*)pData, size, 0);
@@ -316,7 +316,7 @@ namespace BrainCloud
         {
 
             // Read packet
-            int received = ::recv(m_socket, (char*)(m_buffer + m_read), m_maxPacketSize - m_read, 0);
+            size_t received = ::recv(m_socket, (char*)(m_buffer + m_read), m_maxPacketSize - m_read, 0);
             if (received <= 0)
             {
                 FD_CLR(m_socket, &m_fdsSet);
